Preliminary variants and unification through Maude
This is incomplete: terms are not parsed back into correct akiss terms,
which so far has no impact on tests. Old tests run as before and the
most basic XOR test diverges.

Changed the structure of Cime interface to share some code with Maude
interface. Simplified the signature of Variants to omit stuff that did
not make sense in the Maude replacement.