Optim: normalize identical statements for xor
Introduced a new command line option (-xor) that enables normalization
of identical statements: the only identical atoms generated are of
the form identical(R+R',0). This is only a first implem and may not be
fully justified, but it brings a big improvement: rfid0h goes from 1m5
to 20s.