Drop non-normal clauses after canonization
They are useless anyway since their descendants are going to be
dropped by the other non-normality test. But those non-normal
clauses can cause recipization to fail, because it currently
relies on matching modulo the full theory (only AC would be
enough, but it was simpler to implement it this way).