Compute variants through Tamarin
This requires the branch feature-ac-rewrite-rules of tamarin-prover.
Maude's (incomplete) variants are not used anymore.

The new variants require the "deep" rule X+(X+Y)=Y in the case of xor.
For safety it is added (along with other xor rules) as soon as #set xor
is passed, hence we can remove all xor rules from the xor api files.