Reorganize toplevel structure
Break Main into Theory and Main. The first module is responsible for setting
the global parameters concerning the theory, which are then fixed for the
following modules, notably Horn and Main. Main is then responsible for
registering the processes and executing queries.

Eventually this architecture should help us merge Akiss AC with vanilla Akiss
in a clean way.