Use old toolbox for non-AC theories
Move AC-incompatible rewriting toolbox from Term and Variants to Rewriting.
Create Theory.R which uses Tamarin and Maude if AC if required by the theory
or specified on the command line, and the old toolbox otherwise.

Minor tweak in Horn for eqrefl_opt, to handle variables missing from the
substitution.