Fixes on conseq regarding the plus clause
The test rigid.api shows that full conseq leads to inconsistent results.
Updated it for the new version, and added it to the test suite.

In some cases the test failed even with conseq_plus=false. Fix this by
improving is_plus_clause which failed to recognize some permutations
of the plus clause.

Add a new tweak to allow conseq on the plus clause, except when we're
checking a clause that looks like an extension. With this tweak rigid.api
passes even with conseq_plus=true.