Clarify and slightly enhance clause freshening
Freshen before running conseq, to avoid having the same variables in the
clause being checked for conseq and in clauses from the kb. Maude seems to
handle that case nicely, but better be safe.

Additionnally, fine tune where freshening occurs: by doing it before
normalization we benefit more from the "canonizing" effect of the latter
transformation, which identifies more clauses and leads to slightly smaller
knowledge bases.

This modification does not seem to have a performance impact on make test.
In some cases it generates one less tester.