More robust ordering of clauses in Equation
Do not rely on the id of clauses, but check if one is the plus clause.
Tests still run and it does not seem to make a difference.

Also remove unused parameter ?rx'. There is no point to it anyway since
we now forbid equation between two clauses with plus recipes.